tablet_mode: ensure that tablet mode is always initialized
The board init should ensure that tablet_mode is always set so the rest
of the subsystems can rely on this. We assume notebook mode if we don't
have any input from the GSR 360 hall sensor. That matches the default up
in AP firmware as well.

BRANCH=none
BUG=b:141494453,chromium:1010343
BUG=b:160981170
TEST=put arcada at 200 degrees and see that tablet mode it initialized
correctly.

Change-Id: I4e23d3ba149f9add84f9667a5af676803cf50da5
Signed-off-by: Jett Rink <jettrink@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/1845779
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2294633
Tested-by: Marco Chen <marcochen@chromium.org>
Reviewed-by: Marco Chen <marcochen@chromium.org>
Commit-Queue: Marco Chen <marcochen@chromium.org>