chgstv2: stop charging when battery temp exceeds specs
Add a check for the battery temperature in the safe range to charge
the battery, and if it isn't, stop charging. The battery information
defines minimum and maximum temperatures for discharge and for charge.
The state machine already checks if the battery temperature is outside
of the range for discharge (as part of is_battery_critical) and will
shut down the system completely if the battery temperature is out of
range. However, the temperature range for charging is usually tigheter
than for discharging, and it can be safe to discharge, but unsafe to
charge. For example, Kohaku specifies a maximum charge temperature of
55 C, and a maximum discharge temperature of 60 C. If the battery is
at 57 C, we don't want to charge, but it's still OK to use the system.
The check is enabled by CONFIG_BATTERY_CHECK_CHARGE_TEMP_LIMITS.

BUG=b:140596424
BRANCH=None
TEST=`make buildall -j` builds with no errors.
No boards have enabled CONFIG_BATTERY_CHECK_CHARGE_TEMP_LIMITS, so
no effect.

Change-Id: I3b76eab942ca3ef3871f0909395e91634db5640e
Signed-off-by: Paul Fagerburg <pfagerburg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/1914510
Reviewed-by: Scott Collyer <scollyer@chromium.org>
(cherry picked from commit 5e182ed5f641d42dc43fe2c897bf91d47f6470f8)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/1918989
Reviewed-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Commit-Queue: Tim Wawrzynczak <twawrzynczak@chromium.org>
Tested-by: Tim Wawrzynczak <twawrzynczak@chromium.org>