testing: remove incorrect testing assert
When we go through the suspend code path, we disabled RX
monitoring, and we have done that for a 4+ years. We have
not had a unit test for that ever. One is come that needs this

BRANCH=none
BUG=b:144369187
TEST=See that disabling RX in set_state no longer causes
assertion failures in tests that it shouldn't.

Change-Id: Iab4b44d3f5fdd1fe8657b23ac59df247a384ee32
Signed-off-by: Jett Rink <jettrink@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/1925667
Tested-by: Denis Brockus <dbrockus@chromium.org>
Reviewed-by: Denis Brockus <dbrockus@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/1979547
Tested-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Reviewed-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Commit-Queue: Tim Wawrzynczak <twawrzynczak@chromium.org>