ps8xxx: add support for the ps8815
this adds support for the ps8815 variant of the parade TCPC.  this
chip is very similar to its predecessors like the ps8751 and ps8805
and can be supported by the same driver.

at this point, the TCPM can talk to the chip but we don't properly
detect chargers - the CC line states seem wrong and CC status changes
do not trigger an ALERT in the ps8815.

BRANCH=none
BUG=b:144397088,b:147459088
TEST=EC detects the chip on boot.

Change-Id: If86abd1fa21cf8f33f28c4ce89050b29e9408532
Signed-off-by: Caveh Jalali <caveh@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/1969524
Reviewed-by: Daisuke Nojiri <dnojiri@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2392236
Tested-by: Devin Lu <Devin.Lu@quantatw.com>
Commit-Queue: Edward Hill <ecgh@chromium.org>