Kodama: limit the max input current as 2A
The 45W type-c adapter can support 5V/3A and 9V/3A but can not meet DOE
efficiency for those power, so it only put 5V/2A and 9V/2A on the label.
So we limit the input current to 2A to keep same with the certification.

BUG=b:146464104
TEST=the max charger IARCR value is 2A.
BRANCH=kukui

Change-Id: I50b27ecab50b1ff7267afbe80d2f3acf8fff1aba
Signed-off-by: Xiong Huang <xiong.huang@bitland.corp-partner.google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/1984147
Reviewed-by: Eric Yilun Lin <yllin@chromium.org>
Commit-Queue: Eric Yilun Lin <yllin@chromium.org>