ps8xxx: Add workaround for PS8751 FW ver 0x44.
When the AP is powered on, DRP mode is enabled.  Right as the AP is
powered off, and thus DRP mode is disabled, it appears that sometimes
the CC_STATUS register may return incorrect values for CC2 State and/or
CC1 State.  At this point in the process, if there is no port partner
for the TCPC, reading CC_STATUS sometimes returns SNK.Power1.5 instead
of SNK.Open (because no port partner is present).

BUG=b:147472779, b:147684491
BRANCH=firmware-hatch-12672.B
TEST=Repeat reproduction procedure 20 times,
ensure it cannot be reproduced.  Also added
debug code to ensure that the CC_STATUS is
read correctly on the second time around.

Change-Id: I1844b577b314c1c8c2740e7d29aa296821996b63
Signed-off-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2001135
Reviewed-by: Aseda Aboagye <aaboagye@chromium.org>
Reviewed-by: Denis Brockus <dbrockus@chromium.org>
(cherry picked from commit 95a7db7ac0ab2029218a18161315439e853d0fe8)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2002952