cr50: Defer call to board_reboot_ec() in key_combo0_irq()
On Wilco devices the recovery key combination will trigger an irq which
will reboot the EC to enter recovery mode. CL:1881752 moved a 30 ms
delay and made it synchronous. This change works in most contexts except
for within an irq which cannot sleep so when a user presses the recovery
key combo the H1 will hard reboot when the sleep is called. This change
defers the call to board_reboot_ec() so that it is no longer in the irq
context.

BUG=b:147404780
TEST=On Drallion and Sarien press pwr + F2, recovery mode is entered
     without to hold the keys for an extended period of time. Also
     verified that programming flash over CCD still works on Drallion.
BRANCH=none

Change-Id: Ib7ebbe2c67b575b6a99d01a4055bb00ee9d15b7e
Signed-off-by: Mathew King <mathewk@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2020328
Reviewed-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>