common: mag_cal: Trigger online calibration with more sample
For the oneline calibration to work, sensor data must be collected
while the device is in different positions.
The current implementation starts calibration after 25 samples/1 second.
It means the device should done a figure 8 - or close to it - in that
lapse of time. This is fine for phones or small devices, but not for
heavy convertible.

Increase the batch size to at least 50 samples or 2 seconds of
collection.

BUG=b:144027014
BRANCH=reef
TEST=Pass RVCVXCheckTestActivity verifier test.

Change-Id: I78d1b943c23eaa9a29831ad4344c8be36ea00b79
Signed-off-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2034672