chip/stm32/usb: Detect SOF when resume state is not expected
When the resume interrupt handler is called, but the state is not
the expected 0 or 1, let's detect if we receive any SOF from the
host. If we do, this means that the link is active and that we
shouldn't go back to sleep (which would put us in an infinite
loop).

BRANCH=kukui
BUG=b:144808697
TEST=Type usbresume in console (with other HACK CL), see that we
     can recover whether or not the interface is autosuspended.
TEST=With device autosuspended, short D+/D- on the base, that
     generates a spurious resume event (similar to what we
     observed doing ESD testing on Soraka), and check that in that
     can we don't get an SOF and we go back to sleep.

Change-Id: Ibd5c5f7e5f5c324622df5169bb317afa118edc45
Signed-off-by: Nicolas Boichat <drinkcat@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2035439