USB PD: reset VCONN and keep consistent with the previous explicit contract
In one case, TCPC was being a sink but a VCONN source in a previous
explicit contract. TCPC was still be a sink but not go on suppling
VCONN after rebooting EC.
Guarding the saved VCONN role and If TCPC was being sourcing VCONN
and as a sink source in previous explicit contract, TCPC should still
be the VCONN source.

BUG=b:150112192
TEST=Some dongles (PS176-01 and JCA374) can display after
     rebooting EC with adapter.
BRANCH=kukui

Signed-off-by: Xiong Huang <xiong.huang@bitland.corp-partner.google.com>

Change-Id: Ia1e226d21e3a20411521b9ce6f19f749ac48bf51
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2076820
Reviewed-by: Xiong Huang <xiong.huang@bitland.corp-partner.google.com>
Reviewed-by: Ting Shen <phoenixshen@chromium.org>
Reviewed-by: Eric Yilun Lin <yllin@chromium.org>
Tested-by: Xiong Huang <xiong.huang@bitland.corp-partner.google.com>
Commit-Queue: Eric Yilun Lin <yllin@chromium.org>