Make acroterm work with Cr50
One of the difference between Acropora and Cr50 is such that in case
of Acropora the format strings are transferred within packets, and in
case of Cr50 only the index of the string is passed in the packet, the
string itself is obtained from the list prepared by running
util_precompile.py.

Another difference is that in Acropora each console packet message
represents a full line, in Cr50 the message could be representing just
a part of the line, down to a single character.

These differences require modifications of the packet header on the
DUT side and modifications of the Packet class in Acroterm.

Cr50Packet class inherits from Packet, to abstract header structure
differences a class specific method of unpack_ph() is being added. The
new method extracts packet fields, returning fields common for both
packet formats, and saving the class unique values in the object
attributes.

This allows to consolidate the packet validation function and only
divert processing when decode_packet() is invoked. Cr50Packet
constructor is passed the list of format strings, the str_index field
from packet header is the index of the format string in the list.

The Cr50Term class inherits from Acroterm, the only difference is that
Cr50Term uses Cr50Packet for packet processing, and prepares the list
of strings, retrieving it from the blob created by util_precompile.py.

Two new command line options are being added:
 --cr50_mode enables Cr50 packet mode console support
 --cr50_str_blob points at the strings blob, by default
   build/cr50/RW/str_blob is used.

Note that on the DUT the Cr50 packet mode is disabled by default. To
start using console packet one needs to add the line

to board/cr50/board.h and rebuild the image.

BUG=b:149964350
TEST=./util/test_acroterm.py succeeds.
   When Cr50 is built with the suggested modification of
   board/cr50/board.h, acroterm invoked as

   $ util/acroterm.py --tty /dev/ttyUSBx --cr50_mode

   allows to open terminal sessions with USB and UART consoles.

Change-Id: I301a5515a0994dba91f2cb40a77c4b59c3becd45
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2116277
Reviewed-by: Randall Spangler <rspangler@chromium.org>