cmsg: use two byte ints to send function name string indices
When a string parameter __func__ ie encountered in the source code,
the function name is saved in the format strings dictionary, and then
packet is prepared, instead of sending the string, a byte of 0xff is
sent and then the four byte value which is the string index.

But two bytes is enough to send the string index, as it is a 16 bit
value by design.

This patch modifies both transmit and receive sides to start using 2
byte indices for __func__.

BUG=b:149964350
TEST=built the new image, tried running it, observed correct function
     names in the console output.

Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Change-Id: I682dd18cb4dd434e6982d33f1918ef398d5caa20
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2131046
Reviewed-by: Randall Spangler <rspangler@chromium.org>