crypto_api: use const void
This is a minor API clean up, it is not entirely clear why const void
pointers were not used originally, but using this type for input data
(and void pointer for output) makes interfacing with the library much
easier.

Also modified cases where the first parameter of DCRYPTO_SHA1_hash()
was typecasted unnecessarily.

BUG=none
TEST=make buildall succeeds, Cr50 image supports booting a Chrome OS
     device just fine.

Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Change-Id: Ic8a670aa7b26598ea323182845c184b7f1d715a1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2163568
Reviewed-by: Andrey Pronin <apronin@chromium.org>