sps: do not generate AP_INT_L pulse until controller is ready
The recent interrupt pulse extension brought to fore an old bug where
the AP_INT_L pulse is generated by the SPS driver before the
controller is actually ready to accept the next SPI frame.

This patch rearranges the code to make sure that the pulse is
generated after all controller clean up.

BUG=b:154458891
TEST=verified that Atlas device is still booting fine. Will test it on
     other devices which seemed to be triggering the bug.

Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Change-Id: I480760b4afea24295f96abde2fc75c414017c27f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2171452
Reviewed-by: Namyoon Woo <namyoon@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>