ap_ro_integrity_check: allow gsctool erasing AP RO hash
This patch extends the VENDOR_CC_SEED_AP_RO_CHECK handler to erase the
AP RO hash space if two requirements are met:

- the vendor command payload is empty
- the board ID space in INFO1 is not programmed

Also, after this patch it would be impossible to program the AP RO
hash if the Board ID INFO1 field is set. This will prevent attempts to
write the hash by the users of existing devices.

BUG=b:153764696
TEST=after expanding gsctool was able to verify AP RO hash space erase
     when allowed. Was able to write the hash when board ID space is
     uninitialized, and was not able to write the hash when the Board
     ID space is set.

Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Change-Id: I0d2409cb0a97bf98f52e7f10fd41660305638122
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2204975
Reviewed-by: Namyoon Woo <namyoon@chromium.org>