NVMEM: selective erase of TPM objects
This patch adds another NVMEM API, which allows to erase stored TPM
objects selectively. The list of indices of the objects to be erases
is supplied in a zero terminated array.

The existing nvmem_erase_tpm_data() has been modified to erase only
selected objects, if the list of objects is supplied by the caller.

BUG=b:138578447

TEST=Using tpm_manager_client created a bogus NVMEM object, modified
     Cr50 code to provide a CLI command which would invoke the new
     NVMEM API function to delete the new object.

     Invoked 'dump_nvmem' command before and after deleting the bogus
     object. Observed the NVMEM contents compacted and the bogus
     object deleted. Rebooted the device, observed proper Chrome OS
     start up maintaining the existing user account.

Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Change-Id: I3e299c8004141fa01ff20c290131b6526575c42e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2253324
Reviewed-by: Vadim Sukhomlinov <sukhomlinov@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>
Commit-Queue: Andrey Pronin <apronin@chromium.org>