u2f: Append hmac of auth time secret to versioned KH
When generating versioned KHs, u2fd should send a public derivative
(sha256) of the user's auth time secret to cr50. Cr50 derives an
hmac of it and appends this authorization_hmac to the KH.

When signing versioned KHs, u2fd may supply the unhashed auth time
secret. Cr50 will check the authorization_hmac if no power button press.
If the reconstructed hmac matches authorization_hmac, power button press
is waived.

Currently for v1, we will just prepare the authorization_hmac but not
enforce it. This is because fingerprint and PIN are unable to unlock
the same secret.

While we waive power button press for v1, we can enforce
authorization_hmac whenever auth-time secrets is ready.

BUG=b:144861739
TEST=- Use a known 32-byte "auth-time secret"
     - Compute the sha256 of the auth-time secret (this is public)
     - u2f_generate with the computed "authTimeSecretHash"
     - Add code to u2f_sign command handler such that cr50 computes
       the sha256 of the supplied auth-time secret at u2f_sign time
       and require power button press if the hmac doesn't match.
     - u2f_sign with the true auth-time secret -> observe in logging
       that hmac matches, and no power button press required.
     - u2f_sign with a wrong auth-time secret -> observe in logging
       that hmac doesn't match, and power button press is required
       for signing.

Cq-Depend: chromium:2321731
Change-Id: Ib9ae913667f8178ac7a4790f861d7dada972c4a0
Signed-off-by: Yicheng Li <yichengli@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2317047
Reviewed-by: Andrey Pronin <apronin@chromium.org>
Reviewed-by: Vadim Sukhomlinov <sukhomlinov@chromium.org>