g build.mk: fix awk program bug
The recent modification of chip/g/build.mk introduced a bug where an
awk script error is reported if the board name does not contain an
underscore.

It went unnoticed because it does not prevent make from reporting
success and most boards do not care about the results of running of
this awk script.

This patch fixes the problem.

BUG=none
TEST=ran make for cr50 and hslt_d2c, observed proper modification of
     the manifest, verified that no script errors are reported any
     more.

Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Change-Id: I0981411ec7bc17e4473d4b33125f76b75983d974
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2317059
Reviewed-by: Namyoon Woo <namyoon@chromium.org>