driver: bmi160: Read only the first 2 bytes of the interrupt field
The remaining 2 are qualifier for single/double tap we are not
interested in. They are not reset to 0 even after we process the
interrupt, so when gesture is enabled, we are stuck in the interrupt
routine for a while.

BMI260 does not have this problem as the interrupt register is now
explicitly 16 bit.

Gesture is currently enabled on eve that does not have
the loop on interrupt, so no other branch than ToT is affected.

BUG=b:164974014
BRANCH=none
TEST=Without, we would be stuck in bmi160 irq_hanlder:

                (timestamp of int)   (interrupt mask)
...
[5064.237873 a hw int: 769270447 - 2 0x00C00000]
[5064.238989 a hw int: 769270447 - 2 0x00C04000]
[5064.241277 a hw int: 769270447 - 2 0x00C00000]
[5064.242397 a hw int: 769270447 - 2 0x00C00000]
[5064.243527 a hw int: 769270447 - 2 0x00C00000]
[5064.244647 a hw int: 769270447 - 2 0x00C00000]
[5064.246419 a hw int: 769270447 - 2 0x00C00000]
...                                     /
                        Qualifier -----/

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2370152
Reviewed-by: Patryk Duda <pdk@semihalf.com>
Reviewed-by: Daisuke Nojiri <dnojiri@chromium.org>
Tested-by: Patryk Duda <pdk@semihalf.com>
(cherry picked from commit bee0b6334399a341e7e242b5e8b12575262c9791)
Change-Id: I2c8fd354eddfa412f644555dcdcdb77708a9e3c7
Signed-off-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2492532
Reviewed-by: Ching-Kang Yen <chingkang@chromium.org>
Reviewed-by: Yuval Peress <peress@chromium.org>