ln9310: Move battery cell type to board customization
Checking the input voltage to make the decision of 2S battery or 3S
battery is not reliable. The 3S battery with very low charge might be
under 10V that we initialized the switchcap to 2:1 mode. When the
battery was then charged up to 12V, VOUT would go to 6 V which would
be a problem.

This change moves the decision to board customization. It checks the
SKU ID. It can be extended to other ways, like checking the battery
manufacturer and device names.

To prevent any damage when attaching a 3S battery to a 2S-SKU board,
add a check before configuring the switchcap to 2:1 mode.

BRANCH=None
BUG=b:163867792, b:151393598
TEST=Built successfully.

Change-Id: I3f69132bc00b13ec39b229e98a34a5f7f75008f1
Signed-off-by: Wai-Hong Tam <waihong@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2391210
Reviewed-by: Stephen Boyd <swboyd@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2396696
Reviewed-by: Edward Hill <ecgh@chromium.org>
Commit-Queue: Edward Hill <ecgh@chromium.org>
Tested-by: Edward Hill <ecgh@chromium.org>