charge_manager: two charger DUT does not choose correct port
The active charge_port is selected in a delayed call to
charge_manager_refresh. This leaves a big window allowing
a better charger to be detected but not acted upon as
the soon to be active charge_port. This leads to the PD
Policy Engine deciding this better port is not the current
active charge_port and down grading the current limit for
the port to no longer be the best choice.  So when the
charge_manager_refresh is called, it starts changing ports
and then later checks for the best port and then swaps
back to a lesser charger.

BUG=b:166228326
BRANCH=none
TEST=USB-C0 45W then attach USB-C1 65W charger

Signed-off-by: Denis Brockus <dbrockus@google.com>
Change-Id: I23dad1e4690285b3867bd046d9416db9522002e8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2393863
Commit-Queue: Denis Brockus <dbrockus@chromium.org>
Commit-Queue: Edward Hill <ecgh@chromium.org>
Tested-by: Denis Brockus <dbrockus@chromium.org>
Reviewed-by: Edward Hill <ecgh@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2401221
Tested-by: Edward Hill <ecgh@chromium.org>