ec: remove "sushi" variant.
Sushi is not a real product, just a test of the new_variant program.
The effort to keep it up-to-date with the rest of Hatch is no longer
worth it. Remove the variant.

BUG=b:168030592
BRANCH=None
TEST=hatch-cq builds successfully

Signed-off-by: Paul Fagerburg <pfagerburg@chromium.org>
Cq-Depend: chrome-internal:3259091, chromium:2398781
Change-Id: I161bff8ba529b94aaaa9a119155cc2962c686d36
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2399199
Reviewed-by: Furquan Shaikh <furquan@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2406443
Reviewed-by: Edward Hill <ecgh@chromium.org>
Commit-Queue: Edward Hill <ecgh@chromium.org>
Tested-by: Edward Hill <ecgh@chromium.org>