TCPMv2: set object position in SVDM response header
Per PD 6.4.4 about Object Position,
on receiving Enter Mode request, the responder shall set the offset of
mode entered in Object Position field, and it shall be the same value
as that in request;
on receiving Exit Mode request, the responder shall set the offset of
the mode to be exited in Object Position field, and it shall be the same
value as that in request.
This field shall be set to zero in the request or response(REQ, ACK, NAK
or BUSY) when not required by specification of the individual command.

BUG=b:148528713,b:157163664
BRANCH=none
TEST=make buildall -j 4

Signed-off-by: li feng <li1.feng@intel.com>
Change-Id: Icf68d003abb9f1ab9f71fb0646c5e3fb1c5f644a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2399775
Reviewed-by: Abe Levkoy <alevkoy@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2422406
Reviewed-by: Edward Hill <ecgh@chromium.org>
Commit-Queue: Edward Hill <ecgh@chromium.org>
Tested-by: Edward Hill <ecgh@chromium.org>