TCPMv2: Block enabling PD in RO for non-EFS2 and dead-battery
If enable PD in RO on a non-EFS2 device, a hard reset will be issued
when sysjump to RW that makes the device brownout on the dead-battery
case. For this special case, disable PD in RO as a workaround.

BRANCH=None
BUG=b:166307654,b:168134171
TEST=Tested on Pompom, without battery attached. EC warm reset and
sysjump to RW won't cause the board power-lost. EC warm reset may
be brownout but the next boot is fine.

Change-Id: Ie6e7687aa8e160012f9bbd725e0fbcbacae94cf2
Signed-off-by: Wai-Hong Tam <waihong@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2402150
Reviewed-by: Jett Rink <jettrink@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2406577
Reviewed-by: Edward Hill <ecgh@chromium.org>
Commit-Queue: Edward Hill <ecgh@chromium.org>
Tested-by: Edward Hill <ecgh@chromium.org>