Lazor: Fix the battery DSG FET bitmask
When accessing the ManufacturerAccess (0x00) to query the
OperationStatus (0x54), the result is returned from
ManufacturerBlockAccess (0x44) or ManufacturerData (0x23).

Don't need to specify the register of ManufacturerData (0x23)
as it is hard-coded.

In the result of OperationStatus, the DSG FET status is bit-1,
i.e. 0x0002, instead of 0x0001. Fix this bitmask.

The 0x0001 is the status of the system present. Checking the
system present did work for b/159319530, in which the cause
was the long sampling time for the system present. The system
present bit reflects the sampling result.

BRANCH=None
BUG=b:159319530
TEST=Tested on Lazor. Cut-off the battery, plugged AC, and
checked the first boot, no boot loop.

Change-Id: I9289ab6e5d3659f22d239b94ab3b0bdd95693654
Signed-off-by: Wai-Hong Tam <waihong@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2422134
Reviewed-by: Stephen Boyd <swboyd@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2427912
Reviewed-by: Edward Hill <ecgh@chromium.org>
Commit-Queue: Edward Hill <ecgh@chromium.org>
Tested-by: Edward Hill <ecgh@chromium.org>