Trogdor: Enable CONFIG_BOARD_RESET_AFTER_POWER_ON to get correct reset flags
Need this config to save the reset flags across double EC reset on the
first power-on.

On power-on, H1 releases the EC from reset but then quickly asserts and
releases the reset a second time. This means the EC sees 2 resets:
 (1) power-on reset, (2) reset-pin reset. This config will
allow the second reset to be treated as a power-on.

BRANCH=None
BUG=b:169188750
TEST=Tested on a Trogdor board that the first power-on has the
"power-on" reset flag.
> --- UART initialized after reboot ---
> [Image: RO, pompom_v2.0.5526+3b3e20008 2020-09-23 12:00:19 ...]
> [Reset cause: power-on]

Change-Id: I2cccb94d8fac938b4c70984f1a4e4c49b97a8b7b
Signed-off-by: Wai-Hong Tam <waihong@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2427102
Reviewed-by: Stephen Boyd <swboyd@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2427938
Reviewed-by: Edward Hill <ecgh@chromium.org>
Commit-Queue: Edward Hill <ecgh@chromium.org>
Tested-by: Edward Hill <ecgh@chromium.org>