TCPMv2: Report TYPEC commands as a feature
To ease the AP's use of the new TYPEC_* host commands, add a feature
flag to indicate their presence.  Since ToT TCPMv2 always supports these
commands, use that configuration as the trigger for it.

Note for firmware branch pickers: this commit should only be picked if
the branch is also picking the corresponding TYPEC commands (currently
as of this CL: STATUS, DISCOVERY, and CONTROL).

BRANCH=None
BUG=b:167700356
TEST=on waddledoo, confirm Type-c command support shows up in "ectool
inventory"

Signed-off-by: Diana Z <dzigterman@chromium.org>
Change-Id: I4ff78c9b6ca297933611d4abf239fb67b1221751
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2473100
Reviewed-by: Abe Levkoy <alevkoy@chromium.org>
Signed-off-by: Keith Short <keithshort@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2504136