OCPC: Calculate resistance in suspend
Currently our resistance calculations are only made when the system is
charging in S5 or lower.  We can perform these calculations when the
system is in suspend as well since the power consumption is low.
Additionally, it's very possible that a user may not charge in G3 for
a very long time if at all.

BUG=b:170975872
BRANCH=None
TEST=Build and flash waddledoo, suspend DUT, charge from sub board,
verify that resistance is being calculated.

Change-Id: I882eae0014f5c7986e5997872f4a9e004f82ee9d
Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2508851
Reviewed-by: Diana Z <dzigterman@chromium.org>
Commit-Queue: Aseda Aboagye <aaboagye@chromium.org>
Tested-by: Aseda Aboagye <aaboagye@chromium.org>
Signed-off-by: Keith Short <keithshort@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2522136