tcpmv2: ifdef cleanup with IS_ENABLED
Remove as many #if preprocessor statements as possible
without allowing the code to grow when it is compiled.

Verified against the sizes built on current ToT and
this CL on that ToT

~/trunk/src/platform/ec $ make newsizes
Compared 699 of 699 files.
File sizes are unchanged.

BUG=b:172208011
BRANCH=none
TEST=verify TCPMv2 functionality

Signed-off-by: Denis Brockus <dbrockus@google.com>
Change-Id: Ib7d6a41b418f0683d3f7a14236b377073f346010
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2515279
Tested-by: Denis Brockus <dbrockus@chromium.org>
Auto-Submit: Denis Brockus <dbrockus@chromium.org>
Commit-Queue: Jett Rink <jettrink@chromium.org>
Reviewed-by: Jett Rink <jettrink@chromium.org>
Signed-off-by: Keith Short <keithshort@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2522268