ioexpander: Fix type error in command_ioex_get
enum ioex_signal is a uint16. It gets cast to an int32 in
command_ioex_get. When find_ioex_by_name returns -1, it gets
interpreted as 65535, not -1.

Return IOEX_SIGNAL_END from find_ioex_by_name instead of -1.
Minor cleanup to make it more clear that enum ioex_signal is
not an int.

BUG=b:172599383
BRANCH=Zork
TEST='ioexget foo' does not crash

Change-Id: I92922591b6a657a1d2a0840c3d361a8a909895fb
Signed-off-by: Rob Barnes <robbarnes@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2533671
Reviewed-by: Edward Hill <ecgh@chromium.org>
Reviewed-by: Denis Brockus <dbrockus@chromium.org>
Signed-off-by: Keith Short <keithshort@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2541103