zephyr: add i2c_ports enum
This change adds an i2c_ports enum based on the devicetree
phandle list 'named-i2c-ports'. Modeled after named-gpios.
Currently, this is only enabled for ztests as it conflicts
with the definitions in config_chip-npcx7.h. But once that
dependency is removed, this enum will become publicly
available.

BUG=b:171302975
BRANCH=none
TEST=../zephyr-chrome/firmware_builder.py test

Cq-Depend: chromium:2546328
Signed-off-by: Yuval Peress <peress@chromium.org>
Change-Id: I8e763da0c867fed383eb333092f191079fb792ea
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2546997
Reviewed-by: Jett Rink <jettrink@chromium.org>
Signed-off-by: Keith Short <keithshort@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2559339