Dedede: Bail out of hibernate if z-state fails
If we fail to enter z-state, this likely indicates that the z-state
circuitry is in a state from which wake sources will not work.  This
would mean the EC would hibernate until a refresh+power sequence is run.

Instead, choose to reset the system with the AP off.  The battery will
drain more quickly, but will be able to turn on when the power button is
pressed normally.

BRANCH=dedede
BUG=b:166476907
TEST=on drawcia, remove C1 interrupt line sharing and wedge line low.
Ensure "hibernate" run on the EC console from both S0 and G3 results in
a reset with the AP off, and the system can be turned on with a power
button press.  Run with normal C1 interrupt line sharing in place and
verify "hibernate" enters z-state as expected in S0 and G3.

Signed-off-by: Diana Z <dzigterman@chromium.org>
Change-Id: I8408522dd1b0bbbce6f4e2bf6d0c550febd27bbf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2567582
Reviewed-by: Aseda Aboagye <aaboagye@chromium.org>
Commit-Queue: Aseda Aboagye <aaboagye@chromium.org>
Signed-off-by: Keith Short <keithshort@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2570304