sm5803: Add linear charge support
The SM5803 supports manipulating the BFET in the linear region.  This
will be used for the precharge portion of the charge curve in OCPC
systems.  This commit adds support for the linear charge feature for
the charger IC.

BUG=b:174683659
BRANCH=dedede
TEST=With other changes, build and flash drawcia, charge from subboard
on  depleted battery and verify that linear charging is taking place
according to precharge current limits.

Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Change-Id: Ib3c7f14e127b89301cccc6e93848da9249232678
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2570938
Commit-Queue: Aseda Aboagye <aaboagye@chromium.org>
Tested-by: Aseda Aboagye <aaboagye@chromium.org>
Reviewed-by: Diana Z <dzigterman@chromium.org>
Signed-off-by: Abe Levkoy <alevkoy@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2590805