SYV682: Disable VCONN on VCONN OCP
The SYV682 has a 'soft' overcurrent protection on VCONN which current
limits to ~600mA. If the cause is a short it will eventually hit thermal
shutdown. Instead, we should disable VCONN when we get an OCP event. We
must allow current spikes for at least 1ms, so disable with a deglitch.

BUG=b:170441866,b:172710638
TEST=make buildall
TEST=On Voxel, short VCONN to ground for at least 100ms, then observe
that VCONN is disabled and 'VCONN OC!' is printed in the EC console.
BRANCH=None

Signed-off-by: Eric Herrmann <eherrmann@chromium.org>
Change-Id: I6712543b73072c959597c73cd493f248267a42ae
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2572237
Reviewed-by: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>
Signed-off-by: Abe Levkoy <alevkoy@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2580267