battery_fuel_gauge: Make the default battery type customizable
Currently the default battery type is a constant.

There is a need that a board can support both 2S and 3S batteries,
according to its SKU ID. For example, a SKU which uses 2S battery
will use a 2-to-1 switching capacitor to regulate the output
voltage. A SKU which uses 3S battery will use a 3-to-1 switching
capacitor.

If the battery is not attached, the default battery is selected to
configure the charger. A wrong battery configuration may be
selected. For example, the charger may output a 2S voltage but the
switching capacitor uses the 3-to-1 ratio.

This change enables the board to customize the default battery
type. It helps fixing the above issue.

BRANCH=Trogdor
BUG=b:175625362
TEST=Build the Lazor image and use it on the board with 3S battery
but the battery is not attached.

Signed-off-by: Wai-Hong Tam <waihong@google.com>
Change-Id: I76bf50c7226c4ec9f633d4b7f6025dc9ab464b49
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2597801
Reviewed-by: Diana Z <dzigterman@chromium.org>
(cherry picked from commit f55f1f0a5c94d536788d457579c1863e12bc5b0a)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2598193