make: fix no change remake behavior
This patch was inspired by crrev.com/c/2593881, it is extended to
cover both cryptoc and tpm2 libraries. Cherry-picking across the
branches is pointless as the branches have drifted apart and the patch
applied to completely different files.

The below was copied from the original patch description and edited
and augmented to match the Cr50 case.

Currently, the cryptoc and tpm2 targets are marked as
.PHONY. That means that any project that depends on cryptoc or tpm2
will always rebuild ec.bin on make invocations.

For example, running make for cr50 will show the following build steps
on each make invocation, even though nothing in cryptoc or tpm2
changed:

 make obj=/mnt/host/source/src/platform/cr50/build/cr50/cryptoc \
  SUPPORT_UNALIGNED=1 \
    CONFIG_UPTO_SHA512=y -C /mnt/host/source/src/third_party/cryptoc
 make obj=/mnt/host/source/src/platform/cr50/build/cr50/tpm2 \
    EMBEDDED_MODE=1 \
   -C /mnt/host/source/src/third_party/tpm2 copied_objs
 make[1]: Entering directory '/mnt/host/source/src/third_party/cryptoc'
 make[1]: '/mnt/host/.../build/cr50/cryptoc/libcryptoc.a' is up to date.
 make[1]: Leaving directory '/mnt/host/source/src/third_party/cryptoc'
 make[1]: Entering directory '/mnt/host/source/src/third_party/tpm2'
 make[1]: Nothing to be done for 'copied_objs'.
 make[1]: Leaving directory '/mnt/host/source/src/third_party/tpm2'
   LD      RO/ec.RO.elf
   LD      RW/ec.RW.elf
   .
   .
   .

This fix brings the dirty/clean state of cryptoc and tpm2 into the
main Cr50 make process, so that it can assess if libcryptoc.a or tpm
objects (and later ec.bin) actually need to be remade.

We do something similar for the ec version header file that is generated
by the build system itself (see crrev.com/c/227211)

Instead of relying on shell globbing, the tpm2 Makefile is now used to
determine the exact set of the tpm object files to be linked in.

This change was only possible with the fix to cryptoc's Makefile
crrev.com/c/2091999 and to the tpm2 Makefile in crrev.com/c/2606746.

With this change, building Cr50 does not force an unnecessary
recompilation, so the above make example looks like the following
after the initial build:
 $ make BOARD=cr50 -j
 *** 8044 bytes in flash...
 *** 14116 bytes in flash...

BUG=none
TEST=verified that cryptolib and tpm2 libraries are rebuilt when some
     .c or .h file in the respective directories it touched, but not
     unnecessarily, like before this patch.

Change-Id: Ic7c55e6f779559e082afdd18c7368e5115afabdf
Signed-off-by: Vadim Bendebury <vbendeb@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2606810
Tested-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-by: Craig Hesling <hesling@chromium.org>
Commit-Queue: Vadim Bendebury <vbendeb@chromium.org>