TCPMv2: PRL: Fix ordering dependency of rch and tx PRL state machines
With existing PRL architecture, there is a race condition for chunked
messages where the transmit complete flag for the next chunk request
and the next chunk received flags can be both be set. Since the
expectation is that the transmit will be complete before the next
chunk is received, this can cause a false error condition.

This CL modifies the rx chunked state machine to accept the next chunk
when it's expecting a tx complete message as this implies the next
chunk request message was properly sent.

BUG=b:181179550,b:181333342
BRANCH=None
TEST=Verified that chunked message tests passed on the GRL tester. Prior
to this change, chunked message tests would fail on GRL as the tester
responds to the chunk request with little delay.
The tests that this change fixed are:
TD.PD.SNK3.E11
TD.PD.SRC3.E25.

Signed-off-by: Scott Collyer <scollyer@google.com>
Change-Id: I7a7c193952ec5047b9d1f01dec896257274ea96b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2660810
Tested-by: Scott Collyer <scollyer@chromium.org>
Reviewed-by: Abe Levkoy <alevkoy@chromium.org>
Reviewed-by: Diana Z <dzigterman@chromium.org>
Commit-Queue: Scott Collyer <scollyer@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2750036
Tested-by: Abe Levkoy <alevkoy@chromium.org>
Commit-Queue: Abe Levkoy <alevkoy@chromium.org>