TCPMv2: Don't cflush in CC_OPEN with battery
In the TC state machine, on entry to CC_OPEN we disable vbus/vconn,
print a message, flush the console, then disable CC.

But this takes too long to meet the tProtErrHardReset, which causes some
compliance failures. The main culprit is the console flush. The
justification for the console flush is that we may brown out, so this
isn't necessary when we have a battery.

Only do the console flush on entry to CC_OPEN when we don't have a
battery.

BUG=b:181053528, b:196936523
TEST=Compliance test TD.PD.FRSISNK3.E3 on Voxel
BRANCH=None

Signed-off-by: Eric Herrmann <eherrmann@chromium.org>
Change-Id: Ia056f3c3010737e18c02a7a80659da201cfe5f92
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2730630
Reviewed-by: Diana Z <dzigterman@chromium.org>
(cherry picked from commit 698193fa224863758f2fa74828b10773b295b38c)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3116357
Reviewed-by: Denis Brockus <dbrockus@chromium.org>
Commit-Queue: Daisuke Nojiri <dnojiri@chromium.org>
Tested-by: Daisuke Nojiri <dnojiri@chromium.org>