TCPMv2: PRL: Enahnce support for the tx message discard path
This CL modifies the tx message discard path in the PRL. The states
that should have triggered tx message discard events, would only act
on messages pending from PE, but not passed yet to the phy. This
behavior was required to account for the serial nature of the
different PRL state machines and inherent delay in sending/receiving
messages to/from off chip TCPCs via the I2C bus.

This CL adds a new WAIT value to the xmit_status enum so that a
message being sent to the phy but not reponse yet can be known. In
addition, xmit_status == DISCARDED is no longer treated automatically
as a protocol error.

BUG=b:181179550,b:173028144
BRANCH=None
TEST=Running VDMU.E16 compliance test. Validated that the SVDM command
sequence is properly interrupted in most all cases. This test only
fails if collision events are present which is a happening on honybuns
for 2 reasons:
   1) GRL test waits too long to interrupt the VDM command (~1 msec)
   2) honeybuns ucpd is faster since it doesn't have i2c delays.

Signed-off-by: Scott Collyer <scollyer@google.com>
Change-Id: I2bec09b96d7f6ea6122fcb369246717950987ca8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2752251
Tested-by: Scott Collyer <scollyer@chromium.org>
Reviewed-by: Diana Z <dzigterman@chromium.org>
Commit-Queue: Scott Collyer <scollyer@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2799918
Tested-by: Abe Levkoy <alevkoy@chromium.org>
Reviewed-by: Abe Levkoy <alevkoy@chromium.org>
Commit-Queue: Abe Levkoy <alevkoy@chromium.org>