common: motion_sense: Prevent on/off body from waking up device
Introduce a new flag: BYPASS_FIFO to raise an interrupt to the AP
when an event is sent.
It is not as strong as WAKEUP, as it does not wake up the AP.
BYPASS_FIFO must be set when WAKEUP is set.

It is applied to sync sensor as well as gesture/activity sensor.

BUG=b:123434029
BRANCH=trogdor
TEST=Compile

Signed-off-by: Gwendal Grignou <gwendal@chromium.org>
Change-Id: Id5aba71e6d11ef31f7fa2c67ba2e07178d088b1b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2780834
Reviewed-by: Ching-Kang Yen <chingkang@chromium.org>
Tested-by: Ching-Kang Yen <chingkang@chromium.org>
Commit-Queue: Ching-Kang Yen <chingkang@chromium.org>
(cherry picked from commit 4c84afb1cdb1d1da8155f1bc30ea93ecd7b8dd4b)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2793741