remove serialno command
The cr50 serial number is based on the devid, so the serial number can
be found using the sysinfo devid output. The serial number is the devid
without '0x's, capitalized, and the space is replaced with a '-'.

echo ${DEVID//0x/} | tr '[:lower:]' '[:upper:]' | tr ' ' '-'

No scripts or tests use serialno.

This command saves 208 bytes.

BUG=b:181999426
TEST=none

Change-Id: Ib6629ceb9cfa001d8c758077119717f29116cb15
Signed-off-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2799442
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>