boldar: Reset SLP VW signals on EC-AP eSPI reset
This enables an option for the NPCX eSPI support code to reset SLP
signals when eSPI# is asserted because of a global reset.  This is
needed exclusively on platforms that do not use deep Sx.

The EC console apshutdown command forces the AP into G3 but the SLP_S4
signal is not reset in this case as virtual wire signals are not reset
on ungraceful AP shutdown (global reset).

BRANCH=none
BUG=none
TEST=Use EC "powerindebug" command to observe the SLP_S4
	virtual wire is now reset after "apshutdown" command.

Signed-off-by: Sooraj Govindan <sooraj.govindan@intel.com>
Change-Id: I09b78492d24ddeb2e372b7a2390854bd991488a6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2803999
Reviewed-by: caveh jalali <caveh@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2810479
Tested-by: Eric Yilun Lin <yllin@chromium.org>
Reviewed-by: Eric Yilun Lin <yllin@chromium.org>
Commit-Queue: Eric Yilun Lin <yllin@chromium.org>