zephyr: doc: Add a title to the shimming doc
This is confusing when browsing since it has no title. Add one.

Also move the 'alternatives considered' to the bottom while we are
here, since that is confusing for people who are less interested in
the motivation for shimming and just want to know what to do.

BUG=b:181253613
BRANCH=none
TEST=check it in gitiles

Signed-off-by: Simon Glass <sjg@chromium.org>
Change-Id: I73a4716278396dfdf0661d6ced67b3d6a9785165
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2807487
Reviewed-by: Yuval Peress <peress@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2810495
Tested-by: Eric Yilun Lin <yllin@chromium.org>
Reviewed-by: Eric Yilun Lin <yllin@chromium.org>
Commit-Queue: Eric Yilun Lin <yllin@chromium.org>